pyproject.toml
[project]
name = "fastapi-setup"
version = "0.1.0"
description = "A starter template for FastAPI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]",
    "sqlalchemy",
    "asyncpg",
    "alembic",
    "pwdlib[argon2]",
    "pyjwt",
    "greenlet",
    "fastapi-mail",
    "slowapi",
    "structlog",
    "gunicorn",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
]

[tool.fastapi]
entrypoint = "app.main:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
